Move tests for create_trial
